Move logic to module init adding execution capability to __main__ and pycense directly
